Moved and renamed firstpage to event_select and made own dart file for better organization.
